Create confirmation in deleting idea

+++ app/assets/javascripts/delete-idea.js
@@ -6,21 +6,23 @@ $(function() {
     function removeIdea(idea) {
     $('.content__main__memos-area__memo-area[data-id=' + idea.id +']').remove();
     }
-    
+
     var id = $(this).parent().data('id');
 
-    alert('このアイデアを削除します。よろしいですか？');
+    if (confirm('このアイデアを削除します。よろしいですか？')) {
+      $.ajax({
+        type: 'DELETE',
+        url: '/ideas/' + id + '.json'
+      })
+      .done(function(idea) {
+        removeIdea(idea);
+      })
+      .fail(function() {
+        alert('削除に失敗しました。');
+      })
+    } else {
+    }
+      ;
 
-
-    $.ajax({
-      type: 'DELETE',
-      url: '/ideas/' + id + '.json'
-    })
-    .done(function(idea) {
-      removeIdea(idea);
-    })
-    .fail(function() {
-      alert('削除に失敗しました。');
-    })
   })
 })
